Init cleaning of not used components created to test
